+++ rust/psibase_macros/psibase-macros-lib/Cargo.toml
@@ -6,6 +6,6 @@ edition = "2021"
 [dependencies]
 darling = "0.20.10"
 proc-macro2 = "1.0.36"
-syn = { version = "1.0.86" }
+syn = { version = "2.0.79" }
 quote = "1.0.15"
 proc-macro-error = "1.0.4"
